Offer to re-initialize the list of projects/files in org-publish-file.
To save time, we don't re-initialize the list of projects/files each
time a file is published.  But if people want to publish a file that
they just created, this file is not part of any project until the
list of project/files has been re-initialized.

If the file is not part of a project and should be, then the user
can tell so.