Selective tag inheritance.
Tags can now be selected for inheritance by using a list or a regular
expression.

Also for property inheritance, we now allow a regular
expression (list selection existed before).

The property API can now retrieve the value of a property
with inheritance based on the value of
`org-use-property-inheritance'.  Earlier this value was ignored
when querying an entry.  It still is ignored by default, but by
passing `selective' as the INHERIT flag, it can now be used.