Implement description lists.
These are implemented as part of plain lists, so a leading
"-" or "+" is still needed.  Maybe this is not the most pretty
implementation, but it works without rewriting the quite
complex list formatting code.  So all the list formatting
continues to work just fine.  The description tag
is highlighted in bold, so still easy to spot.

The HTML exporter does export these already, the LaTeX
exporter still needs to be adapted.