quickstart edits and regular nomenclature
As it turns out the preamble doesn't have a name in the org syntax
document, so I picked one. This can change if there is a preferred
nomeclature.

I changed the use of the term heading to headline since that is the
term that is used everywhere in the documentation. Might as well get
new users initiated with the correct terminology.