Point to the manual's installation guide
The first and most important part of install.org page was redundant
with the installation instructions on the manual.

Also, the natural expectation is that the manual will always be up to
date, while the page might not be.

The commit promotes the rest of the page in index.org since these are
general information that can be useful to the casual reader.